new perf test

+++ src/test/peregrine/misc/PerfTest2.java
@@ -10,7 +10,7 @@ public class PerfTest2 {
         int range = 1000;
 
         if ( args.length == 2 ) {
-            max = Integer.parseInt( args[0] );
+            max   = Long.parseLong( args[0] );
             range = Integer.parseInt( args[1] );
         }
